+++ test/goldens/skins_opacity/golden_skins_opacity_test.dart
@@ -23,7 +23,7 @@ void main() {
           )!;
           controller.isActive = false;
           artboard.addController(controller);
-          trigger = controller.findInput<bool>('Skin') as SMITrigger;
+          trigger = controller.getTriggerInput('Skin')!;
         },
       );